fix(extension-react-tables): reduce bundle size
