+++ .changeset/clever-baboons-brake.md
@@ -1,0 +1,8 @@
+---
+'@remirror/extension-react-tables': patch
+'@remirror/icons': patch
+'@remirror/react': patch
+'@remirror/react-components': patch
+---
+
+Reduce bundle size by removing `@remirror/icons/all` and `@remirror/react-icons/all-icons` from the package `@remirror/react-tables-extension`.
