+++ packages/remirror__react/src/index.ts
@@ -1,7 +1,6 @@
 export * from '@remirror/extension-react-component';
 export * from '@remirror/extension-react-ssr';
 export * from '@remirror/extension-react-tables';
-export * from '@remirror/extension-react-tables';
 export * from '@remirror/preset-react';
 export * from '@remirror/react-components';
 export * from '@remirror/react-core';
