+++ support/scripts/data/icons.json
@@ -48,6 +48,8 @@
     "delete-bin-fill.svg",
     "pencil-line.svg",
     "pencil-fill.svg",
+    "close-fill.svg",
+    "add-fill.svg",
 
     "a-b.svg",
     "align-bottom.svg",
